pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "evadex"
version = "0.1.0"
description = "Comprehensive DLP evasion test suite — scanner-agnostic, file-aware"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
keywords = ["dlp", "security", "evasion", "testing", "compliance", "pci-dss", "scanner"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "python-docx>=1.1",
    "fpdf2>=2.7.9",
    "openpyxl>=3.1",
    "jinja2>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[project.urls]
Homepage = "https://github.com/tbustenk/evadex"
Repository = "https://github.com/tbustenk/evadex"
"Bug Tracker" = "https://github.com/tbustenk/evadex/issues"
Changelog = "https://github.com/tbustenk/evadex/blob/main/CHANGELOG.md"

[project.scripts]
evadex = "evadex.cli.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
